requirements.txt
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
diff-match-patch==20241021
Django==6.0.1
django-import-export==4.4.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
openpyxl==3.1.5
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
setuptools==80.9.0
sqlparse==0.5.5
tablib==3.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.45.1
yarl==1.22.0
